webrtc/RTCConfiguration-iceCandidatePoolSize.html

<!doctype html>
<meta charset="utf-8">
<!--
4.2.1 RTCConfiguration Dictionary

  The RTCConfiguration defines a set of parameters to configure how the peer to peer communication established via RTCPeerConnection is established or re-established.

  ...

  iceCandidatePoolSize of type octet, defaulting to 0
    Size of the prefetched ICE pool as defined in [JSEP] (section 3.5.4. and section 4.1.1.).
-->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  function initializeWithValidPoolSize(iceCandidatePoolSize, expectedSize) {
    test(() => {
      const pc = new RTCPeerConnection({
        iceCandidatePoolSize
      });
      assert_equals(pc.getConfiguration().iceCandidatePoolSize, expectedSize);

    }, `new RTCPeerConnection({ iceCandidatePoolSize }) with valid value: ${iceCandidatePoolSize}`);
  }

  function initializeWithInvalidPoolSize(iceCandidatePoolSize) {
    test(() => {
      assert_throws(new TypeError(), () => {
        new RTCPeerConnection({
          iceCandidatePoolSize
        });
      });
    }, `new RTCPeerConnection({ iceCandidatePoolSize }) with invalid value: ${iceCandidatePoolSize}`);
  }

  function reconfigureWithValidPoolSize(iceCandidatePoolSize) {
    test(() => {
      const pc = new RTCPeerConnection();
      pc.setConfiguration({ iceCandidatePoolSize });
      assert_equals(pc.getConfiguration().iceCandidatePoolSize, expectedSize);

    }, `setConfiguration({ iceCandidatePoolSize }) to a valid value: ${iceCandidatePoolSize}`);
  }

  function reconfigureWithInvalidPoolSize(iceCandidatePoolSize) {
    test(() => {
      const pc = new RTCPeerConnection();
      assert_throws(new TypeError(), () => {
        pc.setConfiguration({ iceCandidatePoolSize });
      });
    }, `setConfiguration({ iceCandidatePoolSize }) to an invalid value: ${iceCandidatePoolSize}`);
  }

  /*

  dictionary RTCConfiguration {
      ...
      [EnforceRange]
      octet                    iceCandidatePoolSize = 0;
  };

  ... of type octet
  */
  initializeWithValidPoolSize(0, 0);
  initializeWithValidPoolSize(127, 127);
  initializeWithValidPoolSize(255, 255);

  // Defaulting to 0
  initializeWithValidPoolSize(null, 0);
  initializeWithValidPoolSize(undefined, 0);

  // [EnforceRange]
  initializeWithInvalidPoolSize(-1);
  initializeWithInvalidPoolSize(256);

  /*
  setConfiguration({ iceCandidatePoolSize })
  */
  reconfigureWithValidPoolSize(0, 0);
  reconfigureWithValidPoolSize(127, 127);
  reconfigureWithValidPoolSize(255, 255);

  // [EnforceRange]
  reconfigureWithInvalidPoolSize(-1);
  reconfigureWithInvalidPoolSize(256);

</script>
